+++ packages/wallet-sdk/src/components/ConnectContent/ConnectContent.tsx
@@ -4,7 +4,7 @@
 import clsx from 'clsx';
 import { h } from 'preact';
 
-import { createQrUrl } from '../../util';
+import { createQrUrl } from '../../core/util';
 import { LIB_VERSION } from '../../version';
 import { CloseIcon } from '../icons/CloseIcon';
 import coinbaseWalletRound from '../icons/coinbase-wallet-round-svg';
