+++ packages/wallet-sdk/src/core/constants.ts
