+++ packages/wallet-sdk/src/core/errors/constants.ts
