+++ packages/wallet-sdk/src/core/errors/errors.ts
