+++ packages/wallet-sdk/src/core/errors/index.ts
