+++ packages/wallet-sdk/src/core/errors/serialize.ts
@@ -1,5 +1,5 @@
-import { isErrorResponse, Web3Response } from '../relay/Web3Response';
-import { LIB_VERSION } from '../version';
+import { LIB_VERSION } from '../../version';
+import { isErrorResponse, Web3Response } from '../types/Web3Response';
 import { standardErrorCodes } from './constants';
 import { serialize, SerializedEthereumRpcError } from './utils';
