+++ packages/wallet-sdk/src/core/errors/utils.ts
