+++ packages/wallet-sdk/src/core/types/EthereumTransactionParams.ts
@@ -1,6 +1,6 @@
 import BN from 'bn.js';
 
-import { AddressString, IntNumber } from '../types';
+import { AddressString, IntNumber } from '.';
 
 export interface EthereumTransactionParams {
   fromAddress: AddressString;
