+++ packages/wallet-sdk/src/core/types/Web3Provider.ts
@@ -1,7 +1,7 @@
 // Copyright (c) 2018-2023 Coinbase, Inc. <https://www.coinbase.com/>
 // Licensed under the Apache License, version 2.0
 
-import { Callback } from '../types';
+import { Callback } from '.';
 import { JSONRPCRequest, JSONRPCResponse } from './JSONRPC';
 
 export interface Web3Provider {
