+++ packages/wallet-sdk/src/core/util.ts
