+++ packages/wallet-sdk/src/provider/FilterPolyfill.ts
@@ -1,16 +1,16 @@
 // Copyright (c) 2018-2023 Coinbase, Inc. <https://www.coinbase.com/>
 // Licensed under the Apache License, version 2.0
 
-import { HexString, IntNumber } from '../types';
+import { HexString, IntNumber } from '../core/types';
+import { JSONRPCRequest, JSONRPCResponse } from '../core/types/JSONRPC';
+import { Web3Provider } from '../core/types/Web3Provider';
 import {
   ensureHexString,
   hexStringFromIntNumber,
   intNumberFromHexString,
   isHexString,
   range,
-} from '../util';
-import { JSONRPCRequest, JSONRPCResponse } from './JSONRPC';
-import { Web3Provider } from './Web3Provider';
+} from '../core/util';
 
 const TIMEOUT = 5 * 60 * 1000; // 5 minutes
 const JSONRPC_TEMPLATE: { jsonrpc: '2.0'; id: number } = {
