+++ packages/wallet-sdk/src/relay/Session.ts
@@ -3,8 +3,8 @@
 
 import { sha256 } from 'sha.js';
 
+import { randomBytesHex } from '../core/util';
 import { ScopedLocalStorage } from '../lib/ScopedLocalStorage';
-import { randomBytesHex } from '../util';
 
 const STORAGE_KEY_SESSION_ID = 'session:id';
 const STORAGE_KEY_SESSION_SECRET = 'session:secret';
