+++ packages/wallet-sdk/src/relay/WalletSDKRelayEventManager.ts
@@ -1,5 +1,5 @@
-import { prepend0x } from '../util';
-import { Web3Response } from './Web3Response';
+import { Web3Response } from '../core/types/Web3Response';
+import { prepend0x } from '../core/util';
 
 type ResponseCallback = (response: Web3Response) => void;
