+++ packages/wallet-sdk/src/relay/walletlink/ClientMessage.ts
@@ -1,7 +1,7 @@
 // Copyright (c) 2018-2023 Coinbase, Inc. <https://www.coinbase.com/>
 // Licensed under the Apache License, version 2.0
 
-import { IntNumber } from '../types';
+import { IntNumber } from '../../core/types';
 
 export type ClientMessage =
   | {
