+++ packages/wallet-sdk/src/relay/walletlink/DiagnosticLogger.test.ts
