+++ packages/wallet-sdk/src/relay/walletlink/SessionConfig.ts
