+++ packages/wallet-sdk/src/relay/walletlink/WalletLinkConnectionCipher.test.ts
@@ -1,4 +1,4 @@
-import { randomBytesHex } from '../util';
+import { randomBytesHex } from '../../core/util';
 import { WalletLinkConnectionCipher } from './WalletLinkConnectionCipher';
 
 const secret = 'c356fe708ea7bbf7b1cc9ff9813c32772b6e0d16332da4c031ba9ea88be9b5ed';
